requirement.txt
#pip3 install -r requirement.txt
bs4
requests
python-dotenv